Update README.md
Submissions must be a single PDF file named csc384probquiz.pdf, submitted to MarkUs. Submissions must be typed/legible. Please refer to the course information sheet for the late submission policy.

    Bayes’ Burrito Bowls is a food truck that serves burrito bowls (obviously). You can choose:

        a type of rice, which can be; white, brown

        a type of bean, which can be; black, pinto, lima, or kidney

        any of the following toppings; guacamole, corn, salsa, sour cream, onion, lettuce, or cheese

We model each type of burrito as a triple (R, B, T ), where R and B are the types of rice and bean, and